+++ typescript/packages/dash-components/src/WellCompletions/components/Settings/WellPagination.stories.tsx
@@ -1,9 +1,0 @@
-import { exampleDataDecorator } from "../../test/storybookDataDecorator";
-import { withReduxDecorator } from "../../test/storybookReduxAddon";
-import WellPagination from "./WellPagination";
-
-export default {
-    component: WellPagination,
-    title: "WellCompletions/Components/Settings/Well Pagination",
-    decorators: [exampleDataDecorator, withReduxDecorator],
-};