+++ typescript/packages/subsurface-viewer/CHANGELOG.md
@@ -1,3 +1,17 @@
+# [0.3.0](https://github.com/equinor/webviz-subsurface-components/compare/subsurface-viewer@0.2.3...subsurface-viewer@0.3.0) (2023-10-05)
+
+
+### Features
+
+* **grid3d:** constant color for coloring Grid3dLayer ([#1688](https://github.com/equinor/webviz-subsurface-components/issues/1688)) ([718e8a0](https://github.com/equinor/webviz-subsurface-components/commit/718e8a06204fc51c66677415a5bf31ef93860be7))
+
+## [0.2.3](https://github.com/equinor/webviz-subsurface-components/compare/subsurface-viewer@0.2.2...subsurface-viewer@0.2.3) (2023-10-04)
+
+
+### Bug Fixes
+
+* "IsRenderedCallack" shold return true for layers which has visible attribute set to false. ([#1682](https://github.com/equinor/webviz-subsurface-components/issues/1682)) ([816dc37](https://github.com/equinor/webviz-subsurface-components/commit/816dc377d31cc4c8b422124f9e9cb921669afd28))
+
 ## [0.2.2](https://github.com/equinor/webviz-subsurface-components/compare/subsurface-viewer@0.2.1...subsurface-viewer@0.2.2) (2023-09-29)